+++ opt_einsum_fx/_efficient_shape_prop.py
@@ -51,23 +51,47 @@ class EfficientShapeProp(torch.fx.Interpreter):
     during its exploration.)
     """
 
+    def __init__(
+        self, module, garbage_collect_values: bool = True, autocast_default_type=None,
+    ):
+        super().__init__(module, garbage_collect_values=garbage_collect_values)
+        self.autocast_default_type = autocast_default_type
+
+    def _check_dtypes(self, dtypes):
+        dtype_set = set(dtypes)
+        if self.autocast_default_type is not None:
+            assert 1 <= len(dtype_set) <= 2
+            return self.autocast_default_type
+        else:
+            assert len(dtype_set) == 1
+            return next(iter(dtype_set))
+
     def run_node(self, n: Node) -> Any:
         if n.op == "call_function" and n.target in _EINSUM_FUNCS:
             args, kwargs = self.fetch_args_kwargs_from_env(n)
             equation, *operands = args
             shapes = [op.shape for op in operands]
 
-            assert len({op.dtype for op in operands}) == 1
-            meta = SimpleMeta(einsum_shape(equation, *shapes), operands[0].dtype)
-            result = torch.zeros((1,) * len(meta.shape), dtype=meta.dtype, device=operands[0].device).expand(meta.shape)
+            dtype_out = self._check_dtypes(op.dtype for op in operands)
+            meta = SimpleMeta(einsum_shape(equation, *shapes), dtype_out)
+            result = torch.zeros(
+                (1,) * len(meta.shape), dtype=meta.dtype, device=operands[0].device
+            ).expand(meta.shape)
         elif n.op == "call_function" and n.target == torch.tensordot:
             args, kwargs = self.fetch_args_kwargs_from_env(n)
-            shape_a = [dim for i, dim in enumerate(args[0].shape) if i not in kwargs['dims'][0]]
-            shape_b = [dim for i, dim in enumerate(args[1].shape) if i not in kwargs['dims'][1]]
+            shape_a = [
+                dim for i, dim in enumerate(args[0].shape) if i not in kwargs["dims"][0]
+            ]
+            shape_b = [
+                dim for i, dim in enumerate(args[1].shape) if i not in kwargs["dims"][1]
+            ]
 
-            assert len({op.dtype for op in args}) == 1
-            meta = SimpleMeta(shape_a + shape_b, args[0].dtype)
-            result = torch.zeros((1,) * len(meta.shape), dtype=meta.dtype, device=args[0].device).expand(meta.shape)
+            dtype_out = self._check_dtypes(op.dtype for op in args)
+
+            meta = SimpleMeta(shape_a + shape_b, dtype_out)
+            result = torch.zeros(
+                (1,) * len(meta.shape), dtype=meta.dtype, device=args[0].device
+            ).expand(meta.shape)
         else:
             result = super().run_node(n)
 
@@ -77,8 +101,8 @@ class EfficientShapeProp(torch.fx.Interpreter):
                 meta = None
 
         n.meta = dict()
-        n.meta['tensor_meta'] = meta
-        n.meta['type'] = type(result)
+        n.meta["tensor_meta"] = meta
+        n.meta["type"] = type(result)
 
         return result
 
@@ -95,13 +119,13 @@ def einsum_shape(subscripts, *shapes):
        subscripts: the einsum formula
        shapes: the input shapes
     """
-    Shaped = NamedTuple('Shaped', [('shape', tuple)])
+    Shaped = NamedTuple("Shaped", [("shape", tuple)])
     input_subscripts, output_subscript, _ = opt_einsum.parser.parse_einsum_input(
         (subscripts,) + tuple(Shaped(shape) for shape in shapes)
     )
     dims = {
         i: dim
-        for ii, shape in zip(input_subscripts.split(','), shapes)
+        for ii, shape in zip(input_subscripts.split(","), shapes)
         for i, dim in zip(ii, shape)
     }
     return tuple(dims[i] for i in output_subscript)